requirements.txt
streamlit==1.31.0
databricks-sql-connector==3.0.0
python-dotenv==1.0.0
